[api] Adapt escape tests + move it to unit tests (#1661)
